Move the themes classes to a different package
